Fix/general fix

+++ src/services/category.service.ts
@@ -1,4 +1,5 @@
 import Category from "../models/Category";
+import Product from "../models/Product";
 
 export const createCategory = async (data: any) => {
   const existingCategory = await Category.findOne({ name: data.name });
@@ -16,7 +17,7 @@ export const getCategories = async () => {
   return category;
 };
 
-export const getCategoryById = async (id: String) => {
+export const getCategoryById = async (id: string) => {
   const category = await Category.findById(id);
 
   if (!category) {
@@ -33,15 +34,32 @@ export const updateCategory = async (id: string, data: any) => {
     throw new Error("Catgory Not Found");
   }
 
+  if (data.status) {
+    await Product.updateMany(
+      { categoryId: id },
+      { $set: { status: data.status } },
+    );
+  }
+
   return category;
 };
 
 export const deleteCategory = async (id: string) => {
-  const category = await Category.findByIdAndDelete(id);
+  const category = await Category.findById(id);
 
   if (!category) {
     throw new Error("Category Not Found");
   }
+
+  const hasProduct = await Product.exists({ categoryId: id });
+
+  if (hasProduct) {
+    throw new Error(
+      "Cannot Delete Category Because It Is Assing To One or More Products",
+    );
+  }
+
+  await Category.findByIdAndDelete(id);
 
   return { message: "Category Deleted Successfully" };
 };
